- Fixed overflow when sum_hits >= 65536, causing pR = NaN. NaN result   reported by Florian Schlichting <fschlich@CIS.FU-Berlin.DE> for   osbf-lua-2.0.4.

+++ osbf_bayes.c
@@ -808,9 +808,9 @@ void osbf_bayes_classify(const unsigned char *p_text,   /* pointer to text */
           if (cfx > 1)
             cfx = 1;
           confidence_factor = cfx *
-              pow((diff_hits * diff_hits - K1 /
+              pow(((double)diff_hits * diff_hits - K1 /
                    (classes[i_max_p]->hits + classes[i_min_p]->hits)) /
-                  (sum_hits * sum_hits), 2) /
+                  ((double)sum_hits * sum_hits), 2) /
               (1.0 +
                K3 / ((classes[i_max_p]->hits + classes[i_min_p]->hits) *
                      feature_weight[window_idx]));
